+++ exportplan/core/helpers.py
@@ -1,6 +1,4 @@
 import pytz
-from django.conf import settings
-from hashids import Hashids
 from iso3166 import countries_by_alpha3
 
 from core import models
@@ -138,26 +136,3 @@ def get_exportplan(sso_session_id, id):
     response = api_client.exportplan.detail(sso_session_id=sso_session_id, id=id)
     response.raise_for_status()
     return response.json()
-
-
-hashids = Hashids(settings.HASHIDS_SALT, min_length=8)
-
-
-def h_encrypt(id):
-    return hashids.encrypt(id)
-
-
-def h_decrypt(h):
-    z = hashids.decrypt(h)
-    if z:
-        return z[0]
-
-
-class HashIdConverter:
-    regex = '[a-zA-Z0-9]{8,}'
-
-    def to_python(self, value):
-        return h_decrypt(value)
-
-    def to_url(self, value):
-        return h_encrypt(value)
